+++ nifti2bids/parsers.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-from typing import Optional
 
 import pandas as pd
 
@@ -19,7 +18,7 @@ PRESENTATION_COLUMNS = [
 ]
 
 
-def _determine_deliminator(textlines: list[str], column_headers: list[str]) -> str:
+def _determine_deliminator(textlines: list[str]) -> str | None:
     """
     Identify the deliminator used for the data based on the
     deliminator used for the column titles.
@@ -29,18 +28,15 @@ def _determine_deliminator(textlines: list[str], column_headers: list[str]) -> s
     textlines: :obj:`list[str]`
         The lines of text from the presentation log file.
 
-    column_headers: :obj:`list[str]` or :obj:`None`
-        The column headers for the data in the Presentation log file.
-
     Returns
     -------
-    str
-        The deliminator
+    str or None
+        The deliminator or None if the deliminator not determined.
     """
     for line in textlines:
-        if line.startswith(column_headers[0]):
-            split_text = line.split(column_headers[1])[0]
-            deliminator = split_text.removeprefix(column_headers[0])
+        if line.startswith(PRESENTATION_COLUMNS[0]):
+            first_string = line.split(PRESENTATION_COLUMNS[1])[0]
+            deliminator = first_string.removeprefix(PRESENTATION_COLUMNS[0])
             break
 
     return deliminator
@@ -73,8 +69,42 @@ def _convert_textlines_to_df(
     return pd.DataFrame(data, columns=column_headers)
 
 
+def _convert_time(
+    presentation_df: pd.DataFrame, convert_to_seconds: bool
+) -> pd.DataFrame:
+    """
+    Convert timing of the Presentation Dataframe to floats.
+
+    Parameters
+    ----------
+    presentation_df: :obj:`DataFrame`
+        Pandas Dataframe of the Presentation log
+
+    convert_to_seconds: :obj:`bool`, default=False
+        Convert resolution of all time columns from 0.1ms to seconds.
+
+    Returns
+    -------
+    pandas.Dataframe
+        Dataframe of Presentation log with timing converted to floats and
+        time resolution converted to seconfs if ``convert_to_seconds`` is
+        True.
+    """
+    # Convert timing from strings to floats
+    columns = set(["Time", "TTime", "Duration", "ReqTime", "ReqDur"])
+    present_columns = list(columns.intersection(presentation_df.columns))
+    presentation_df[present_columns] = presentation_df[present_columns].astype(float)
+
+    if convert_to_seconds:
+        presentation_df[present_columns] = presentation_df[present_columns].apply(
+            lambda x: x / 10000
+        )
+
+    return presentation_df
+
+
 def load_presentation_log(
-    log_filepath: str | Path, column_headers: Optional[list[str]] = None
+    log_filepath: str | Path, convert_to_seconds: bool = False
 ) -> pd.DataFrame:
     """
     Loads Presentation log file as a Pandas Dataframe.
@@ -84,42 +114,23 @@ def load_presentation_log(
     log_filepath: :obj:`str` or :obj:`Path`
         Absolute path to the Presentation log file (i.e text, log, excel files).
 
-    column_headers: :obj:`list[str]` or :obj:`None`, default=None
-        The column headers for the data in the Presentation log file.
-        If None, then the following headers are used:
-
-        ::
-
-            default_column_headers = [
-                "Trial",
-                "Event Type",
-                "Code", "Time",
-                "TTime",
-                "Uncertainty",
-                "Duration",
-                "Uncertainty",
-                "ReqTime",
-                "ReqDur",
-                "Stim Type",
-                "Pair Index"
-                ]
-
+    convert_to_seconds: :obj:`bool`, default=False
+        Convert resolution of all time columns from 0.1ms to seconds.
 
     Returns
     -------
-    Dataframe
+    pandas.Dataframe
         A Pandas dataframe of the data.
     """
     with open(log_filepath, "r") as f:
-        column_headers = column_headers if column_headers else PRESENTATION_COLUMNS
         textlines = f.readlines()
-        deliminator = _determine_deliminator(textlines, column_headers)
+        deliminator = _determine_deliminator(textlines)
         content_indices = []
 
         cleaned_textlines = [line for line in textlines if line != "\n"]
         for indx, line in enumerate(cleaned_textlines):
             # Get the starting index of the data columns
-            if line.startswith(f"{deliminator}".join(column_headers)):
+            if line.startswith(f"{deliminator}".join(PRESENTATION_COLUMNS)):
                 content_indices.append(indx)
             # Get one more than the final index of the data colums
             # Note: the lines for the data contain a trial number
@@ -134,4 +145,6 @@ def load_presentation_log(
 
         data_textlines = cleaned_textlines[(start_indx + 1) : stop_indx]
 
-    return _convert_textlines_to_df(data_textlines, deliminator, column_headers)
+    df = _convert_textlines_to_df(data_textlines, deliminator, PRESENTATION_COLUMNS)
+
+    return _convert_time(df, convert_to_seconds)