+++ tests/test_bids.py
@@ -7,7 +7,9 @@ from nifti2bids.bids import (
     create_participant_tsv,
     create_dataset_description,
     save_dataset_description,
+    presentation_log_to_bids,
 )
+from ._constants import BLOCK_DATA, EVENT_DATA
 
 
 @pytest.mark.parametrize("dst_dir, remove_src_file", ([None, True], [True, False]))
@@ -71,3 +73,57 @@ def test_create_participant_tsv(tmp_dir):
 
     df = pd.read_csv(filename, sep="\t")
     assert df["participant_id"].values[0] == "sub-01"
+
+
+def _create_presentation_logfile(dst_dir, design):
+    dst_dir = Path(dst_dir)
+    data = BLOCK_DATA if design == "block" else EVENT_DATA
+
+    filename = dst_dir / f"{design}.txt"
+    with open(filename, mode="w") as file:
+        for line in data:
+            file.write("\t".join(line) + "\n")
+
+
+@pytest.mark.parametrize("experimental_design", ("block", "event"))
+def test_presentation_log_to_bids(tmp_dir, experimental_design):
+    """Test for ``presentation_log_to_bids``."""
+    from pandas.testing import assert_frame_equal
+
+    filename = Path(tmp_dir.name) / f"{experimental_design}.txt"
+    _create_presentation_logfile(tmp_dir.name, experimental_design)
+
+    if experimental_design == "block":
+        expected_df = pd.DataFrame(
+            {
+                "onset": [1.0, 35.0],
+                "duration": [14.0, 14.0],
+                "trial_type": ["indoor", "indoor"],
+            }
+        )
+        kwargs = {
+            "presentation_log_or_df": filename,
+            "trial_types": ["indoor"],
+            "experimental_design": "block",
+            "rest_block_code": "rest",
+        }
+    else:
+        expected_df = pd.DataFrame(
+            {
+                "onset": [7.9107, 10.8965],
+                "duration": [0.7058, 0.6720],
+                "trial_type": ["incongruentright", "neutralleft"],
+            }
+        )
+        kwargs = {
+            "presentation_log_or_df": filename,
+            "trial_types": ["incongruent", "neutral"],
+            "experimental_design": "event",
+        }
+
+    df = presentation_log_to_bids(**kwargs)
+    assert_frame_equal(df, expected_df)
+
+    if experimental_design == "event":
+        expected_df["response"] = ["hit", "hit"]
+        df = presentation_log_to_bids(**kwargs, include_response=True)